added circles to cover up the objects
I added some circles to cover the objects since I want to make a game. The player has to use the flashlight to find hidden objects. The black circle will disappear when the mouse touches it. The circles are black for now so I can clearly see where they are, but I will make them the same color as the background.