finished conditionals for one cover
At first I wanted to make the cover move when it overlapped with the flashlight. However, after the flashlight moved, the bear was still visible, which is not what I wanted. I tried changing the size of the cover but the problem was the same. Then used the changed the color of the cover to a transparent when it overlapped with the flashlight and when it was not overlapping, I made the cover go back to it's original color.