made the sky change colors
I made the sky have a night and day cycle to make it more challenging for players. I also added a rectangle over the whole canvas so that the everything is harder to see, giving the illusion that it is nighttime. The rectangle is invisible during the day and gets darker at night