added more colors to the frog, more flies and another bird
I added another fly and another bird to make the game more interesting. When the tongue overlaps with the bird, the frog will start dying and change colors. There are 3 colors depending on how close to the frog is to death which acts as a warning to the player. The frog slowly heals after a few second. The second bird appears a few second after the game starts to make the game more challenging.