added menu
I added the instructions to start the game. The instruction will disappear when space is pressed and the start time will also start when the instructions disappear to indicate the start of the game. Because of this I changed the speed for the color change (for the sky) and made the second bird appear 5 seconds after the start time instead of 7 seconds.