adjusted the time to make bird2 appear and disappear
I adjusted the time for bird2 to appear and disappear because it was appearing before the instructions were gone and then not appearing at all when the game started. I reduced the time passed since the start of the game and the opacity required for the bird to disappear so the bird appears for longer. I also made the flashlight appear when the opacity is at 110 instead of 105 so that it does not appear right after the second bird is gone.