added water and changed the size of the canvas
I made a slight change to the size of the canvas because I wanted the water to be more visible without having less space for the sky. I added some water for the frog to swim in. I also changed the size of the black rectangle to make sure everything gets darker.