added a half a second delay before showing the ending
I added a half a second delay before showing the ending after the frog dies. While playing the game, the flashlight would often appear on the frog and the player would die without knowing what happened. I added this to show that the frog did change colors before showing the end so that the player can see that the frog died.