added another way for the frog to die
I needed a way to force the frog to eat flies to make the game more interesting. I made it so that the frog starts dying if it does not eat. Since I only had 3 colors (excluding the original color), I decided to make the frog change colors every 3 seconds. If the frog does not eat for 9 seconds total, it dies and the game ends. The timer starts from when the frog last ate, so it starts off false at the beginning of the game.