added a smaller bird
I added a smaller birds that the frog can eat. Since it is spring, baby birds are born and leave the nest with their mothers to search for food. Since frogs eat small animals, it can eat the smaller birds to heal instantly. However, if the frog eats a small bird while it's healthy, the birds will damage the frog.