made the frog hide underwater
I made the frog hide underwater to heal and hide form the flashlight. I wanted to find a way to make the game different from other version and make the game easier to play. I did not want to add more flies because it would not make sense for the colder weather and I have reused that idea twice already. I wanted to try something new.