added a bird and added another flashlight
I added another bird with a blue beak that will kill the frog if it is eaten. There are a lot of birds in the sky in autumn and I thought it would be interesting to give the frog another predator to make the game more challenging. I also added another flashlight to make the experience more entertaining. Rather than killing the frog right away, both flashlights need to touch the frog to kill it.