added a brown fly
I added a brown fly that can teleport. I wanted the game to be less challenging to win because in the original version of the game, there was no way to win except avoiding the flashlight which was really hard to do. I wanted to make it easier for the player to win the game.